Refactor Terminal functions
